+++ Documento.cs
@@ -9,14 +9,14 @@ namespace csharp_biblioteca_db
     public enum Stato { Disponibile, Prestito }
     public class Documento
     {
-        public int Codice { get; set; }
+        public long Codice { get; set; }
         public string Titolo { get; set; }
         public string Settore { get; set; }
         public Stato Stato { get; set; }
         public List<Autore> Autori { get; set; }
         public Scaffale Scaffale { get; set; }
 
-        public Documento(int Codice, string Titolo, string Settore, string nomeScaffale)
+        public Documento(long Codice, string Titolo, string Settore, string nomeScaffale)
         {
             this.Codice = Codice;
             this.Titolo = Titolo;
