+++ Libro.cs
@@ -10,7 +10,7 @@ namespace csharp_biblioteca_db
     {
         public int NumeroPagine { get; set; }
 
-        public Libro(int Codice, string Titolo, string Settore, int NumeroPagine,string Scaffale) : base(Codice, Titolo, Settore,Scaffale)
+        public Libro(long Codice, string Titolo, string Settore, int NumeroPagine,string Scaffale) : base(Codice, Titolo, Settore,Scaffale)
         {
             this.NumeroPagine = NumeroPagine;
